Move CasLogout to separate file (to match master)
