+++ scala-cas_2.12/src/main/scala/fi/vm/sade/utils/cas/CasClient.scala
@@ -135,14 +135,4 @@ private[cas] object JSessionIdClient {
   }
 }
 
-object CasLogout {
-  def parseTicketFromLogoutRequest(logoutRequest: String): Option[String] = {
-    Utility.trim(XML.loadString(logoutRequest)) match {
-      case <samlp:LogoutRequest><saml:NameID>{nameID}</saml:NameID><samlp:SessionIndex>{ticket}</samlp:SessionIndex></samlp:LogoutRequest> =>
-        Some(ticket.text)
-      case _ => None
-    }
-  }
-}
-
 class CasClientException(message: String) extends RuntimeException(message)
