+++ scala-cas_2.12/src/main/scala/fi/vm/sade/utils/cas/CasLogout.scala
@@ -1,0 +1,14 @@
+package fi.vm.sade.utils.cas
+
+import scala.xml.{Utility, XML}
+
+object CasLogout {
+  def parseTicketFromLogoutRequest(logoutRequest: String): Option[String] = {
+    Utility.trim(XML.loadString(logoutRequest)) match {
+      case <samlp:LogoutRequest><saml:NameID>{nameID}</saml:NameID><samlp:SessionIndex>{ticket}</samlp:SessionIndex></samlp:LogoutRequest> =>
+        Some(ticket.text)
+      case _ => None
+    }
+  }
+}
+
